// components/student/studentMyProject/studentMyProjectEditor/studentMyProjectEditorStyle.scss.d.ts
export type Styles = {
  'areaActionButtonsPreviewImageFile': string;
  'categoryContentContainer': string;
  'categorySeparator': string;
  'errorText': string;
  'imgesContainer': string;
  'inputText': string;
  'itemImgContainer': string;
  'popupBodySize': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;